Add ActivateVolume helper APIs for activating LUKS volumes
This adds 2 functions for activating LUKS volumes.

ActivateVolumeWithTPMSealedKey attempts to activate a volume using the TPM
protected secret. If a PIN is required to unseal the secret, it will be
requested using systemd-ask-password. Once unsealed, the secret will be
used to activate the LUKS volume using systemd-cryptsetup.

ActivateVolumeWithTPMSealedKey can fall back to activating with the
fallback recovery key if activating with the TPM protected secret fails.

ActivateVolumeWithRecoveryKey can be used to activate a volume with the
fallback recovery key directly without attempting first with the TPM
protected secret.

In both functions, the fallback recovery key is requested using
systemd-ask-password. When activation with the recovery key is successful,
the key is added to the user keyring in the kernel. The description of
the keyring entry will contain an indicator of the reason that the
fallback recovery key was used.