Add PCRProtectionProfile.computePCRDigests
Previously, PCRProtectionProfile would compute collections of PCR values
for each branch, and then computeDynamicPolicy would compute a PCR selection
and list of PCR digests from this.

Have PCRProtectionProfile compute the PCR selection and list of PCR digests
instead, as it will make it easier for PCRProtectionProfile to de-duplicate
paths that result in the same PCR values.