Make PCRProtectionProfile.computePCRValues non recursive
AddEFISecureBootPolicyProfile currently operates non-recursively. A
future PR will change the interaction between
AddEFISecureBootPolicyProfile and PCRProtectionProfile so that it
creates a profile using the methods of PCRProtectionProfile rather
than computing PCR values itself and adding them to the profile with
PCRProtectionProfile.AddPCRValue as it does now. This change will
preserve the property that computing a PCR profile for the secure
boot policy is non-recursive.