Refactor the interaction between PCRProtectionProfile and AddEFISecureBootPolicyProfile
PCRProtectionProfile computes PCR values from a sequence of instructions
and has logic to de-duplicate identical branches. That functionality is
duplicated in AddEFISecureBootPolicyProfile, which computes PCR values,
de-duplicates identical values and then adds the computed values to the
PCR profile with a single PCRProtectionProfile.AddPCRValue call for each
branch.

This refactors the interaction between the 2 pieces of code to remove the
duplicated functionality from AddEFISecureBootPolicyProfile. Now, it
builds a profile using sequences of PCRProtectionProfile.ExtendPCR and
PCRProtectionProfile.AddProfileOR calls, and relies on
PCRProtectionProfile to compute the final PCR values and de-duplicate
identical branches.