Fix the signature handling in secureBootPolicyGen.computeAndExtendVerificationMeasurement
The pkcs7 package gives us the signer certificate and we check whether
this is either in the UEFI db or is trusted by a CA certificate in the
UEFI db in order to compute the verification measurement for a signed
binary. This isn't entirely correct though, as the Authenticode signature
can contain a set of intermediate certificates that form a chain, and the
signing certificate doesn't have to be directly signed by the CA
certificate. This fixes that to handle signatures where there are
intermediate certificates between the signer and the CA.