Add support for computing authentication events for dual-signed binaries
This adds support for computing authentication events for binaries that
have multiple signatures. Computing these events relies on determining
which CA certificate will be used to authenticate an image, and this
requires an assumption about how firmware implementations behave.

This implementation assumes that firmware implementations will iterate
over a binary's signatures in the order in which they appear in an outer
loop (ie, for each signature, attempt to authenticate the image using one
of the CA certificates). EDK2 works like this, and the firmware on my
Intel NUC also does.

If implementations iterate over the authorized signature database in an
other loop instead (ie, for each CA certificate, attempt to authenticate
the image using one of its signatures), then this assumption may generate
PCR profiles that are incorrect for dual-signed binaries in some cases.
I've added this limitation (and some others) to the documentation.

This commit also contains a refresh of some test binaries with an updated
signing certificate that contains a different name and serial.