Don't use x509.Certificate.Verify in AddEFISecureBootPolicyProfile
https://github.com/snapcore/secboot/pull/58 converted
AddEFISecureBootPolicyProfile to use x509.Certificate.Verify to detect
chains of trust between an Authenticode signing certificate and a
trusted CA certificate. This is the best way to do it because it enables
the code to detect chains that involve intermediate certificates in the
Authenticode signature. However, UEFI doesn't perform time checking when
validating certificates during EFI image authentication, so expired
certificates are considered valid. Go's crypto/x509 API doesn't allow us
to disable time checking, and so we can't use it here to detect chains of
trust.

The Microsoft certificate used to sign the most recent version of shim
(Microsoft Windows UEFI Driver Publisher) expired on May 2, 2020.

Go back to just checking whether a signing certificate is signed by a
trusted CA certificate, and document the limitations of this.