Add ErrNoTPM2Device and TPMConnection.IsEnabled
This makes use of tpm2.TPMContext.IsTPM2, newly added to go-tpm2 in order
to determine if the TPM device is actually a TPM2 device during connection.
If there is no TPM device or there is a TPM device but it is not a TPM2
device, then a new ErrNoTPM2Device error is returned.

Also added is TPMConnection.IsEnabled, which is used to determine whether
the TPM has been disabled by the platform firmware. A TPM can be disabled
by the platform firmware but still remain visible to the operating system.