Store the TPM simulator's persistent data in a temporary directory
This creates a temporary directory for storing the TPM simulator's
persistent data. For the snap version of the TPM simulator, it uses the
snap's common data directory instead, because it has a private tmpdir in
which we can't create a temporary directory outside of the snap.

Once this code moves in internal/testutil, it will have the ability to
specify a location for the persistent data, so that the simulator can run
with pre-generated, checked-in data.