Test that changes are backwards compatible with supported metadata versions
This adds an initial set of tests to make sure that future changes are
backwards compatible with key files created with the v0 version of the
on-disk metadata. The tests operate on a checked in key file and TPM
simulator state, created with a secboot version that creates key files
with the v0 metadata version.

In the future, additional tests will be added for every revision of the
on-disk metadata format version.

The tests cover operations that are expected to continue working on key
files created with older secboot versions, such as
UpdateKeyPCRProtectionPolicy, ChangePIN and SealedKeyObject.UnsealFromTPM.