Create internal/testutil package to allow multiple packages to share test code
This is required to facilitate the addition of a new set of tests for
making sure we don't regress backwards compatibility.

It's private for now, but we could consider making it public in order to
support tests with the TPM simulator inside snapd.

This also requires the addition of several other internal subpackages, in
order to provide to internal/testutil some of the functionality that was
provided via export_test.go.