Use a named pipe for passing keys to cryptsetup where stdin can't be used
systemd-cryptsetup can't read the key from stdin, so we passed it via a
temporary file. Whilst cryptsetup can read a key from stdin, it can
only read a single key and so isn't suitable for providing both keys in
the luksAddKey case. In this case, we passed both the existing key and
new key via a temporary file, and this couldn't be a named pipe because
it is required to be seekable.

Switch to using a named pipe for passing the key to systemd-cryptsetup.
For "cryptsetup luksAddKey", pass the existing key via a named pipe and
pass the new key on stdin. The "cryptsetup luksFormat" case is unchanged -
the initial key is passed to cryptsetup via stdin.