Don't use a NV index for PIN integration
The current implementation uses a NV index for PIN integration using
the TPM2_PolicySecret assertion. In the original technical plan, PIN
integration was going to be achieved using a separate TPM object (and
TPM2_PolicySecret), although this was changed to a NV index instead.

There are a couple of reasons for this approach:
- The original plan didn't feature the use of the TPM2_PolicyAuthorize
assertion for allowing the PCR policy to be updated without having to
reseal - instead, updating the PCR policy was going to be achieved by
resealing (ie, creating a new sealed key object). Having a separate
object for PIN integration avoids the need to know the PIN in order to
perform a resealing operation.
- I initially thought that the sealed key object containing the disk
encryption key required the userWithAuth attribute in order to use the
object's auth value in its own authorization policy, and we can't set that
attribute on the sealed key object becuase it would mean it could be
unsealed without satisfying the authorization policy.

This last point isn't correct though - a sealed key object's auth value
can be used in its own authorization policy using the TPM2_PolicyAuthValue
assertion without requiring the userWithAuth attribute (it can't be used
in the authorization policy for another object with TPM2_PolicySecret
though).

The current implementation makes use of the TPM2_PolicyAuthorize assertion
so that the PCR policy can be updated without resealing (using
UpdateKeyPCRProtectionPolicy).

This refactors the PIN integration to no longer depend on a NV index,
although the NV index remains because it serves one other purpose (as an
optional counter for revoking PCR policies).