Store the private part of the policy auth key inside the sealed key object
Now that an elliptic curve key is used to authorize PCR policy updates and
this is significantly smaller than the previous RSA key, store the private
part of the key inside the sealed key object and have it returned from
SealedKeyObject.UnsealFromTPM rather than storing it inside a file inside
the encrypted volume.

This removes the need to store an additional file in order to use
UpdateKeyPCRProtectionPolicy.

ActivateVolumeWithTPMSealedKey now adds the private part of the key to
the root user's keyring. Callers to UpdateKeyPCRProtectionPolicy will
need to retrieve the key data from there using the keyctl syscall.