Reimplement the support for LockAccessToSealedKeys to reduce fragility
Access to sealed keys can be denied until the next TPM reset or restart
by calling LockAccessToSealedKeys. The traditional way of achieving this
is by extending another value to a PCR that is included in the PCR
selection for a sealed key's authorization policy after the key has been
unsealed, which makes its authorization policy non-satisfiable until the
next reset or restart. The issue with this is that it makes the feature
dependent on the PCR profile.

The current implementation of LockAccessToSealedKeys uses a NV index at a
well known handle and takes advantage of a couple of properties of NV
index read locks:
- Once enabled, they can only be disabled by a TPM reset or restart.
- Enabling a read lock sets the index's TPMA_NV_READLOCKED attribute which
changes its name.
Sealed keys created by this package have an authorization policy that
asserts that this NV index exists at its well known handle. Calling
LockAccessToSealedKeys enables the read lock for this index, which changes
its name and makes this assertion fail.

The problem with this approach though is that the TPM owner (ie, anyone
who knows the authorization value for the storage hierarchy, which is
empty by default) can undefine and redefine a NV index in order to clear
the read lock attribute, thus re-enabling access to sealed keys. One
approach to prevent this from happening relies on the fact that writing to
a NV index for the first time sets the TPMA_NV_WRITTEN attribute which
also changes its name. If the index is defined with the
TPMA_NV_POLICY_WRITE attribute and an authorization policy that can only
be satisfied by an assertion signed with an ephemeral key (using
TPM2_PolicySigned), and sealed keys have authorization policies that can
only be satisfied by the presence of the initialized index, then it is not
possible for an adversary without the private part of the ephemeral key to
undefine and redefine an identical NV index in order to remove the
TPMA_NV_READLOCKED attribute and reenable access to sealed keys. This is
the approach currently taken by secboot, but this has its own problem - if
the NV index is accidentally undefined, then this would make all sealed
keys permanently unrecoverable. Ideally, provisioning the TPM without
clearing it should be able to remedy any accidental changes and restore
access to valid sealed keys in all cases except where the storage primary
seed has been changed (ie, the TPM has been cleared).

This changes the approach used for protecting the NV index used by
LockAccessToSealedKeys. The new approach makes use of 2 NV indices and has
the property that the NV indices have names that are common to all devices
and can always be recreated, but they have authorization policies that
enforce a creation sequence that means that they can only be created in
the locked state. The way this works is as follows:
- A bootstrap index is created at handle 2 with TPMA_NV_AUTHREAD and
TPMA_NV_AUTHWRITE attributes and an empty authorization policy.
- An empty write is performed to initialize the bootstrap index.
- An index is created at handle 1 with TPMA_NV_AUTHREAD,
TPMA_NV_POLICY_WRITE and TPMA_NV_READ_STCLEAR attributes, and an
authorization policy that can only be satisfied by the presence of the
initialized bootstrap index.
- An empty write is performed to initialize index 1 using a policy session.
- The bootstrap index is undefined and another index is created at handle
2 with TPMA_NV_AUTHREAD and TPMA_NV_POLICY_WRITE attributes, and an
authorization policy that can only be satisfied by the presence of the
initialized index at handle 1 with its read lock enabled.
- The read lock for the index at handle 1 is enabled.
- An empty write is performed to initialize index 2 using a policy session.

Authorization policies for sealed keys must assert the presence of both of
these NV indices using the names of the initialized indices without the
read lock enabled. Enabling the read lock on index 1 will disable access
to these keys.

It isn't possible to recreate these indices in their unlocked state.
Although they can be recovered if they are undefined accidentally, they
will not return to their unlocked state until the next TPM reset or
restart.

An adversary can't undefine and redefine either index in order to
reenable access to sealed keys. Eg, say an adversary performs the
following actions:
- They undefine and redefine index 1. The new index isn't read locked, but
sealed keys can't be unsealed until the new index has been written to. But
it has an authorization policy that requires index 2 to be undefined and
replaced with the bootstrap index.
- They undefine index 2 and create the boostrap index in its place.
- They perform an empty write to index 1 to initialize it. Index 1 now has
the expected name, but sealed keys still can't be accessed because index 2
no longer exists.
- They undefine the bootstrap index again and redefine index 2. However,
sealed keys still can't be unsealed until the new index has been written
to. But it has an authorization policy that requires the read lock to be
enabled on index 1.
- They enable the read lock on index 1.
- They perform an empty write to index 2 to initialize it. Index 2 now has
the expected name, but sealed keys still can't be accessed because index 1
is read locked. This can ony be undone by performing a TPM reset or
restart.