Add GetActivationDataFromKernel
GetActivationDataFromKernel retrieves information added to the user
keyring by one of the ActivateVolume calls. In the case where a volume is
activated with the TPM sealed key, this will return the private part of
the policy authorization key. In the case where a volume is activated with
a recovery key, this will return a structure containing the recovery key
and the reason that the recovery key was requested.