Make AddEFISecureBootPolicyProfile less strict
AddEFISecureBootPolicyProfile tries to determine the
EV_EFI_VARIABLE_AUTHORITY event in PCR7 associated with the verification
of shim, and whether that event was also used for verification of UEFI
drivers (in this case, the event needs to be preserved in the PCR profile,
even if we create a profile where shim is verified with another
authority). But the code for doing this is overcomplicated and too
strict - it finds the first EV_EFI_BOOT_SERVICES_APPLICATION in PCR4
(which is associated with shim), then assumes that the preceeding
EV_EFI_VARIABLE_AUTHORITY event in PCR7 is the one associated with the
verification of it. If this event in PCR7 occurs before the transition to
OS-present (ie, before the EV_SEPARATOR events in PCRs 0-6) then it
assumes that it was also used for verification of a UEFI driver.

This breaks on firmware that measures the EV_EFI_VARIABLE_AUTHORITY and
EV_EFI_BOOT_SERVICES_APPLICATION events in a different order.

Instead, simplify this by preserving all EV_EFI_VARIABLE_AUTHORITY events
in the generated PCR profile if they occur before the transition to
OS-present. Also handle the case where the EV_SEPARATOR event for PCR7 is
measured as part of the transition to OS-present as opposed to being
measured before the hand off to BDS (most implementations measure this
before the handoff to BDS, which is before any UEFI drivers or system
preparation applications are executed).

Fixes https://github.com/snapcore/secboot/issues/107