Add initial platform-agnostic API
This adds an initial implementation of a platform-agnostic key API.
Several new types are added:

- KeyData, which is the main representation of a key.
- KeyDataReader, which is the interface that KeyData uses to read
key data from persistent storage.
- KeyDataWriter, which is the interface that KeyData uses to write
key data to persistent storage.
- KeyCreationData, which contains the information required to create
a new KeyData.
- KeyPayload, which represents the data protected by a KeyData object.
- PlatformKeyData, which represents the data exchanged between this
package and platform implementations.
- PlatformKeyDataHandler, which is the interface used by this package
to communicate with platform implementations.

Key data is stored in JSON.

The new API contains a mechanism to bind the key data to snap models,
so that the early boot environment can restrict access to encrypted
data to a previously authorized model. The binding is by way of a HMAC
created with a key that is stored in the encrypted payload along with
the disk unlock key.