Automatically create a KeyID and make it genuinely unique
Rather than expecting the consumer of the API to supply a key
name which is used to create a KeyID, and rather than storing an
ID inside the KeyData, just make the KeyID automatically generated,
completely unique and make KeyData.ID always return a fresh value.

As the KeyID isn't human readable, also introduce KeyData.ReadableName,
which just returns a string for use in error messages. For file-loaded
KeyData, this is just the file path.