Create tpm2 subpackage
This moves the TPM code out of the core package and into a subpackage
of its own. This allows non-TPM platforms to avoid pulling in all of the
TPM related code.

Note that some code is duplicated between crypt.go + tpm2/crypt.go and
crypt_test.go + tpm2/crypt_test.go. I haven't made the effort to refactor
this so that the code can be shared because I'm going to try to make
it possible to unlock LUKS devices with the current TPM sealed key files
using the new KeyData based API (crypt.go). This will work by constructing a
dummy KeyData with an empty encrypted payload, the platform data set
to the serialized TPM sealed key data, and a platform name of "tpm2-legacy"
or something like that. The "tpm2-legacy" platform implemented in tpm2/ will
take care of the unsealing.

Future PRs will work towards creating a proper tpm2 KeyData platform
implementation whilst maintaining the ability to use existing TPM key files.
A big pre-requisite for this is to remove the SealedKeyObject hard dependency
on file paths so that it only cares about bytes.