Remove ActivateOptions field from ActivateVolumeOptions
This is currently unused by snapd.

The field is problematic because it exposes all of systemd-cryptsetup's
options, many of which are incompatible with how we use it in secboot.

If there is a genuine need in the future to expose some extra options,
this can be done with a more explicit API instead.

This also makes it slightly easier to move the TPM code out of the
core secboot package because it's not necessary to duplicate
makeActivateOptions or move it to an internal package.