Support provisioning the TPM with a custom SRK
TPMConnection.EnsureProvisioned will create and persist a srorage
primary key as described in the "TCG TPM v2.0 Provisioning Guidance"
specification (https://trustedcomputinggroup.org/wp-content/uploads/TCG-TPM-v2.0-Provisioning-Guidance-Published-v1r1.pdf).

This specification says that the unique field of the storage primary
key template should be 256 bytes of zeroes, which is the same as the
endorsement key template.

However, many existing tools (including tpm2-tools) define the template
for the storage primary key with an empty unique field, and this creates
a different primary key. This is a deviation from the TCG specifications.

Whilst this isn't a problem for Ubuntu Core, it may be a problem on
systems where the initial encryption and key protection happens before
first boot, and the disk unlock key is protected with a primary key that
is created with third party tools.

In order to support this scenario, introduce a new
TPMConnection.EnsureProvisionedWithCustomSRK API. It is not expected that
this API will be used on Ubuntu Core, but may be used elsewhere.

Note that SealKeyToTPM unconditionally recreates the storage primary
key if it is executed with a TPMConnection that didn't perform the
provisioning step. The reason for this is that creating the primary
key is really the only way to ensure that we are sealing to a key that
we know we can recreate if it is deleted in the future. In order to
avoid extending the SealKeyToTPM API, the new EnsureProvisionedWithCustomSRK
function will store the custom template used to create the primary key
inside the TPM. SealKeyToTPM will use this template stored inside the TPM
(if it exists) to recreate the storage primary key.