Initial support SBAT-capable shim
This makes efi.AddSecureBootPolicyProfile compatible with a SBAT
capable version of shim, which measures the current value of the SBAT
variable to PCR7. The initial support is very limited - it only supports
the value that is currently compiled into shim and which shim sets the
variable to, but this is ok for now.

Some things to note:
- SBAT-capable shim will initialize the SBAT variable to a known (compiled
in) payload if the variable doesn't exist, has an older payload or doesn't
have the correct attributes. "Older" right now is determined by checking
a timestamp in the payload whilst the variable is BS+NV and
not-authenticated, but would be determined by the firmware using the
signature timestamp for an authenticated variable in the future.
- The variable is initialized as BS+NV and then mirrored to a RT variable
by a SBAT-capable shim, which means we don't know what the current
variable value is if we are pre-computing a PCR policy on a system that
was booted with a pre-SBAT shim. This isn't a problem right now because
there is only one payload, but will be a problem in the future as updates
are published if the variable remains a non-authenticated BS+NV variable
as opposed to being a RT+BS+NV authenticated variable.
- In the future and in order to do this properly, we need an authoritative
source for the current variable value (eg, event log for BS+NV variable or
current variable value for RT+BS+NV, like we do for other configuration).
- Shim doesn't provide a way to audit the compiled-in SBAT payload, so we
don't have a way to introspect what it will set it to, although we know
what it is right now.
- Future PCR value computation will be a bit more complicated than it is
now - imagine if you have 2 shim's with different built-in SBAT payloads,
and those are both different to the current SBAT value. Because shim will
overwrite the SBAT variable if its built-in payload is newer, booting
with one shim may affect the PCR values associated with a branch that has
a different shim.