Add RevokeOldPCRProtectionPolicies API
The UpdatePCRProtectionPolicy API computes a new PCR policy, writes
the updated key data to disk and then revokes old PCR policies.
In order to remove SealedKeyObject's hard dependency on files, the
second step of writing the updated key data to disk has to be dropped,
as this will be handled outside of SealedKeyObject.

This means that the PCR policy revocation must move to its own API,
as revocation has to happen after the updated key data is persisted
to disk.