Big update to EFI tests
The tests in ./efi/ depend on a lot of checked-in binary data,
all of which is either not really reproducible, or only manually
reproducible following steps that are mostly stored in my head.
This makes it difficult to update or add to the testdata when
required - a task that currently can only really be done by me.

This makes several changes:
- Adds a efi/testdata/src directory containing a fairly minimal set
of sources for testdata. This is a mix of public certs, source code
for mock EFI binaries, public uefi.org dbx updates, checked in test
RSA private keys and an eventlog from a Dell box to verify a specific
bug fix.
- Adds tools/make-efi-testdata to (mostly reproducibly) compile
the actual test data in efi/testdata. The compiled testdata is still
checked in because reproducibility of the mock EFI binaries requires
the use of the same toolchain, but it can be updated or re-created
by anyone. This tool creates mock event logs, mock EFI variable stores,
mock EFI binaries, mock signature database updates and creates mock
certificates from the checked-in RSA private keys.
- Updates the tests to use gopkg.in/check.v1, with some updates to
the actual tests performed.

These updates did find one bug - we would compute the wrong digest
where grub and the kernel are both authenticated using a CA stored
in the UEFI authorized signature database with the latest shim,
because the status of a flag was not correctly propagated between
two steps.

This requires updating both go-efilib and tcglog-parser. This required
quite a bit of work in both projects to support creating the test
data.

One limitation of this is that signing EFI binaries and signature
database updates is not quite reproducible because the signatures
contain timestamps. A refresh of the testdata will always create files
with a different hash, but the different file hashes don't affect the
actual tests.