Add internal/argon2 package
This adds an internal package for deriving keys from passphrases
using the Argon2i algorithm, and will be used for passphrase support.

There are 2 core APIs:
- Key is just a wrapper around the argon2.Key API provided by the
golang.org/x/crypto/argon2 package, but using the CostParams type
returned by Benchmark.
- Benchmark determines the cost parameters for the algorithm based
on the supplied target duration, maximum memory cost and key length.
It is based on the algorithm used by cryptsetup, which prioritizes
memory cost over time cost.

The algorithm consumes a lot of memory and isn't suitable for long
lived system processes in a garbage collected language. For the
secboot public APIs that execute argon2, I'll add an interface that
allows the caller to either run the algorithm in-process if
appropriate, or delegate it to a short-lived utility process that can
be defined outside of secboot.

Benchmark runs the algorithm multiple times and can cause the current
process to rapidly run out of memory. To support delegating this to
a short-lived utility process, it provides a callback argument for
running each actual measurement and returning a duration. The
KeyDuration function provides the in-process implementation of this,
and it is expected that this will be executed in a short-lived utility
process using a mechanism defined outside of this package. It is
possible to run the entire benchmark in-process by running the
garbage collector after each measurement, and this is what the unit
tests are doing, but this isn't really appropriate for production
code.