tpm2: Add "legacy" KeyData platform implementation
This adds a KeyData platform implementation to allow
keys to be recovered from the current (soon to be legacy)
TPM sealed key files using the new secboot.KeyData API. This
permits the use of the secboot.ActivateVolumeWithKeyData API
for TPM sealed key files, allowing the code in tpm2/crypt.go
to be deleted.

In order to do this, a new function for reading TPM sealed key
files is introduced - NewKeyDataFromSealedKeyObjectFile, which
constructs a secboot.KeyData from a sealed key file. Note that
the snap model authorization API can't be used for these KeyData
instances, and snap-bootstrap shouldn't use this API for files
opened via it. That's ok though, because the PCR profile contains
an assertion of the snap model.

When the tpm2 package becomes a proper KeyData platform, it's
expected that these newer KeyData's will use the snap model
authorization API like any other platform. We need to be careful
to ensure that the new KeyData can't be manipulated in such a way
to trick snap-bootstrap into reading it via the (what will be
legacy) NewKeyDataFromSealedKeyObjectFile API, recovering the key
successfully via the legacy platform and skipping the snap model check.

The deleted code in tpm2/crypt.go attempted to create a primary key
if the initial unseal failed, before attempting it one more time.
To preserve this behaviour, the code is now part of UnsealFromTPM.