Switch tpm2 tests to github.com/canonical/go-tpm2/testutil
Some of the testing code in this repository is duplicated in
go-tpm2, so it's now provided as a public package.

Some of the functionality provided by go-tpm2/testutil includes:
- A mechanism for starting and shutting down a TPM simulator.
- Fixtures that ensure TPM state is restored at the end of each
test, when the connection is closed (and which fail a test if it
makes changes that cannot be undone by the fixture). This stops
TPM state leaking between tests as long as each test runs with its
own connection, and prevents hidden dependencies between tests
(something which the secboot tests have been relying on).
- Fixtures that sandbox the set of features a test can use and
which can skip tests that require features not permitted by the
current test environment. This permits a subset of tests to be
executed on a real TPM rather than having a hard dependency on
a simulator, by whitelisting certain features via commandline
arguments (eg, use of hierarchies, PCRs, NV storage, the ability
to clear the TPM etc).

I've documented some of go-tpm2/testutil:
https://pkg.go.dev/github.com/canonical/go-tpm2/testutil

This does temporarily increase the amount of code in each test
because of the signature of tpm2test.OpenTPMConnectionT(). I'm going
to rewrite all of the existing tests to not use the testing package
in a future PR, where the connection is managed per-suite rather
than per-test.