internal/luks2: Add support for external token implementations
Only 2 token fields are required for all tokens (type + keyslots),
with remaining fields being type specific. To handle this, tokens
are represented by a struct containing the 2 required fields and
a map[string]interface{} to handle the remaining fields.

Working with a map[string]interface{} type is a pain, so introduce
a way to decode to type-specific token representations instead.
Handlers for these are registered by calling RegisterTokenHandler
from an external package.